+++ src/OutputMatcher.php
@@ -11,7 +11,11 @@ final readonly class OutputMatcher
 
     public function getExpectedOutput(): string
     {
-        if ($this->parsedFile->expectedRegexFile !== "") {
+        if ($this->parsedFile->expectedPatternFile !== "") {
+            return (string) @file_get_contents($this->parsedFile->expectedPatternFile);
+        } elseif ($this->parsedFile->expectedPattern !== "") {
+            return $this->parsedFile->expectedPattern;
+        } elseif ($this->parsedFile->expectedRegexFile !== "") {
             return (string) @file_get_contents($this->parsedFile->expectedRegexFile);
         } elseif ($this->parsedFile->expectedRegex !== "") {
             return $this->parsedFile->expectedRegex;
@@ -36,8 +40,47 @@ final readonly class OutputMatcher
     {
         $expectedOutput = $this->getExpectedOutput();
         return match ($this->getMode()) {
-            OutputMatcherMode::Regex => (bool) preg_match("/^$expectedOutput\$/s", $actualOutput),
+            OutputMatcherMode::Regex, OutputMatcherMode::Substitution => (bool) preg_match((string) $this->getRegex(), $actualOutput),
             default => $actualOutput === $expectedOutput,
         };
+    }
+
+    public function getRegex(): ?string
+    {
+        if ($this->getMode() === OutputMatcherMode::Literal) {
+            return null;
+        } elseif ($this->getMode() === OutputMatcherMode::Regex) {
+            return "/^" . $this->getExpectedOutput() . "\$/s";
+        }
+        $result = $this->getExpectedOutput();
+        $result = str_replace(
+            [
+                "%e",
+                '%s',
+                "%S",
+                "%a",
+                "%A",
+                "%w",
+                "%i",
+                "%d",
+                "%x",
+                "%f",
+                "%c",
+            ],
+            [
+                DIRECTORY_SEPARATOR,
+                "[^\\r\\n]+",
+                "[^\\r\\n]*",
+                ".+",
+                ".*",
+                "\\s*",
+                "[+-]?\\d+",
+                "\\d+",
+                "[0-9a-fA-F]+",
+                ".",
+            ],
+            preg_quote($result, "/")
+        );
+        return "/^" . $result . "\$/s";
     }
 }