+++ tests/PhptRunnerTest.php
@@ -195,5 +195,23 @@ final class PhptRunnerTest extends TestCase
         $this->assertSame(Outcome::Passed, $result->outcome);
         $this->assertSame("test error", $result->output);
         $this->assertSame("test error", $result->expectedOutput);
+
+        $filename = __DIR__ . DIRECTORY_SEPARATOR . "test_substitutions.phpt";
+        $result = $runner->runFile($filename);
+        $this->assertSame($filename, $result->fileName);
+        $this->assertSame("Test substitutions", $result->testName);
+        $this->assertSame("", $result->testDescription);
+        $this->assertSame(Outcome::Passed, $result->outcome);
+        $this->assertSame("+123 abc test", $result->output);
+        $this->assertSame("%i%w%s test", $result->expectedOutput);
+
+        $filename = __DIR__ . DIRECTORY_SEPARATOR . "test_substitutions_external.phpt";
+        $result = $runner->runFile($filename);
+        $this->assertSame($filename, $result->fileName);
+        $this->assertSame("Test substitutions external", $result->testName);
+        $this->assertSame("", $result->testDescription);
+        $this->assertSame(Outcome::Passed, $result->outcome);
+        $this->assertSame("11f test ext", $result->output);
+        $this->assertSame("%x%sext", $result->expectedOutput);
     }
 }