+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Version 0.2.0-dev
 - only first line of sections TEST, ARGS and FILE_EXTERNAL is now parsed
 - sections GET, COOKIE and PHPDBG are parsed now
-- implemented sections EXPECTREGEX, EXPECTREGEX_EXTERNAL and CAPTURE_STDIO
+- implemented sections EXPECTREGEX, EXPECTREGEX_EXTERNAL, CAPTURE_STDIO, EXPECTF and EXPECTF_EXTERNAL
 
 Version 0.1.0
 - initial version